requirements.txt
absl-py==2.0.0
appdirs==1.4.4
arandr==0.1.11
asgiref==3.6.0
attrs==23.2.0
av==10.0.0
Babel==2.10.3
beautifulsoup4==4.11.2
beniget==0.4.1
blinker==1.5
Bottleneck==1.3.5
Brotli==1.0.9
cffi==1.16.0
chardet==5.1.0
click==8.1.3
colorama==0.4.6
colorzero==2.0
configobj==5.0.8
contourpy==1.2.0
cryptography==38.0.4
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.3.2
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
et-xmlfile==1.0.1
filelock==3.13.1
Flask==2.2.2
flatbuffers==20181003210633
fonttools==4.47.2
fs==2.4.16
fsspec==2023.12.2
gast==0.5.2
gitdb==4.0.11
GitPython==3.1.41
gpiozero==2.0
html5lib==1.1
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
jdcal==1.0
Jinja2==3.1.2
json-rpc==1.13.0
kiwisolver==1.4.5
lgpio==0.2.2.0
libevdev==0.5
lxml==4.9.2
lz4==4.0.2+dfsg
MarkupSafe==2.1.2
matplotlib==3.6.3
mediapipe==0.10.9
more-itertools==8.10.0
mpmath==1.3.0
netaddr==0.8.0
netifaces==0.11.0
networkx==3.2.1
numexpr==2.8.4
numpy==1.24.2
odfpy==1.4.2
olefile==0.46
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
openpyxl==3.0.9
packaging==23.2
pandas==2.1.4
pexpect==4.8.0
pgzero==1.2
picamera2==0.3.17
pidng==4.0.9
piexif==1.1.3
pigpio==1.78
pillow==10.2.0
pluggy==1.0.0+repack
ply==3.11
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
py==1.11.0
py-cpuinfo==9.0.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pygame==2.1.2
Pygments==2.14.0
PyGObject==3.42.2
pyinotify==0.9.6
PyOpenGL==3.1.6
pyOpenSSL==23.0.0
pyparsing==3.1.1
PyQt5==5.15.9
PyQt5-sip==12.11.1
pyserial==3.5
pysmbc==1.0.23
pytest==7.2.1
python-apt==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-prctl==1.8.1
pythran==0.11.0
pytz==2022.7.1
pyudev==0.24.0
pywifi==1.1.12
PyYAML==6.0.2
pyzbar==0.1.9
pyzmq==26.4.0
reportlab==3.6.12
RPi.GPIO==0.7.1a4
RTIMULib==7.2.1
scipy==1.11.4
seaborn==0.13.1
sense-hat==2.6.0
simplejpeg==1.6.6
simplejson==3.18.3
six==1.16.0
smbus2==0.4.2
smmap==5.0.1
sounddevice==0.4.6
soupsieve==2.3.2
spidev==3.5
sympy==1.12
tables==3.7.0
terminator==2.1.2
thop==0.1.1.post2209072238
toml==0.10.2
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
typing_extensions==4.13.2
tzdata==2023.4
ufoLib2==0.14.0
ultralytics==8.1.2
v4l2-python3==0.3.3
webencodings==0.5.1
Werkzeug==2.2.2
zipp==1.0.0
